requirements.txt
Django==1.8.1
django-colorfield==0.1.4
django-extensions==1.5.5
django-filter==0.10.0
django-grappelli==2.6.4
djangorestframework==3.1.2
gunicorn==19.3.0
Pillow==2.8.1
psycopg2==2.6
six==1.9.0
Werkzeug==0.10.4
